refactor: Split experience.tsx into focused files
651 lines → 4 files:
- experience-types.ts   — shared ExperienceItem type + constants
- experience-detail.tsx — slide-in detail panel component
- experience-card.tsx   — clickable preview card component
- experience.tsx        — scroll-jack container + public Experience export

No behaviour change.
